update controllers, move the auth flow to an auth controller
