+++ spring-back-end/pjl-authentication/pjl-authentication-core/src/main/java/com/pjlsoftware/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.pjlsoftware.controller.security;
+package com.pjlsoftware.controller;
 
 import com.pjlsoftware.entity.User;
 import com.pjlsoftware.projection.UserProjection;
@@ -6,15 +6,12 @@ import com.pjlsoftware.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/users")
@@ -25,31 +22,6 @@ public class UserController {
     @Autowired
     public UserController(UserRepository userRepository) {
         this.userRepository = userRepository;
-    }
-
-    @RequestMapping(
-            value = "/create/google",
-            method = RequestMethod.POST,
-            produces = {"application/json"}
-    )
-    public ResponseEntity<String> createGoogleUser() {
-        JwtAuthenticationToken jwtAuthenticationToken = (JwtAuthenticationToken) SecurityContextHolder.getContext().getAuthentication();
-        String bearerToken = jwtAuthenticationToken.getToken().getTokenValue();
-
-        User user = ValidateGoogleAuthToken.verifyGoogleAuthToken(bearerToken)
-                .orElseThrow(() -> new RuntimeException("Failed to validate JWT."));
-
-        Optional<User> isAlreadyUser = userRepository.findByUsername(user.getUsername());
-
-        if (isAlreadyUser.isPresent()) {
-            User existingUser = isAlreadyUser.get();
-            existingUser.setEnabled(true);
-            userRepository.saveAndFlush(existingUser);
-            return new ResponseEntity<>("{\"value\": \"Re-enabled existing user\"}", HttpStatus.CREATED);
-        } else {
-            userRepository.saveAndFlush(user);
-            return new ResponseEntity<>("{\"value\": \"Created new user\"}", HttpStatus.CREATED);
-        }
     }
 
     @RequestMapping(
